HW-3 Main의 Repository 부분 MVP 구현
